Added explicit cast and parentheses to gnl_break_str() call in the return of gnl_read(), avoiding implicit cast from pointer to integer, and making clear the intention of breaking a string just after the new line character.

+++ exam_03/gnl/get_next_line.c
@@ -267,7 +267,7 @@ char	*gnl_read(int fd, char **buffer)
 	}
 	if (read_buf)
 		free(read_buf);
-	return (gnl_break_str(buffer, new_line - *buffer + 1));
+	return (gnl_break_str(buffer, (int)(new_line - *buffer) + 1));
 }
 
 /**
